requirements.txt
colorama
pywhatkit
mysql-connector-python
pymongo